requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
psycopg==3.2.5
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
